#11 load photo assets; Data Service refactoring
