+++ EAO/PFPhotoThumb.swift
@@ -20,7 +20,8 @@ class PFPhotoThumb: Object, Mappable{
     
     ///Use this variable for image caching
     var image : UIImage? {
-        return nil
+        let url = URL(fileURLWithPath: FileManager.directory.absoluteString).appendingPathComponent(id, isDirectory: true)
+        return UIImage(contentsOfFile: url.path)
     }
     
     // MARK: Properties
